refactor(kernel): widen kernel-side address types to 64-bit
Stage 3a of the address-space relocation work. Userspace is still i386;
this widens only the kernel-side plumbing so the ring-3 ABI flip (3c) is a
separate, bisectable change.

tcb_t's address fields were uint32_t because userspace was i386-only:
esp, eip, cr3, sig_handler, image_base, image_end, ipc_reply_buf, argv_ptr
and sig_altstack_sp are now uint64_t. priority and cap_tcb are not
addresses and stay 32-bit. This also fixes a latent truncation that was
already live: create_task stores DEMO_TASK_STACK_TOP (0x7fffe0000000 - 256)
into esp, which a uint32_t silently folded to 0xdfffff00.

The Rust FFI is widened to match (rust_validate_page_fault,
rust_get_user_page_protection). rust_get_user_page_protection was already
truncating high stack addresses at its call site in paging.c. Widening the
Rust signature turned that into a compile error rather than a silent
narrowing -- the C-side equivalent had been converting quietly for as long
as it had been wrong.

Two comments in kernel.h claimed image_base and sig_altstack_sp were
"carved from padding so the struct size is unchanged". That is no longer
true (sizeof(tcb_t) 448 -> 488), so the claim is removed rather than left
to mislead. Nothing depends on the layout: the only offsetof assertions in
the tree are on capability_t, and no assembly reaches into tcb.

Two build-correctness fixes, both the same class of bug -- the build
reporting a result that does not reflect the sources:

  -MMD -MP: without header dependencies, editing a header rebuilt nothing
  that included it, and the link reused stale objects. This masked four
  real -Wincompatible-pointer-types errors in kspawn.c behind a "build: 0"
  clean report during this very change. A green build now means something.

  make smoke: alone among the eighteen smoke targets it did not clean
  first, so it booted whatever kernel the previous target left behind --
  `make smoke-newlib && make smoke` tested the NEWLIB_SELFTEST kernel and
  timed out. TESTS.md already documents these targets as clean-building;
  smoke was the outlier. CI never saw it (fresh checkout per job), so it
  only ever misled locally.

Verified: 18/18 smoke targets pass (including smoke-smp), 57 Rust tests,
clippy -D warnings clean, kernel.elf byte-for-byte reproducible, and no
RWX LOAD segment.

+++ rust/src/lib.rs
@@ -42,8 +42,8 @@ impl SafeCap {
 
 /// Low user stack window that `create_user_pagedir` premaps (32 pages below
 /// `0x7ff000`). Fixed, so it is a constant here rather than a per-task bound.
-const LOW_STACK_BASE: u32 = 0x7df000;
-const LOW_STACK_TOP: u32 = 0x7ff000;
+const LOW_STACK_BASE: u64 = 0x7df000;
+const LOW_STACK_TOP: u64 = 0x7ff000;
 
 /// Is `fault_addr` a legitimate part of THIS task's user address space?
 ///
@@ -60,12 +60,12 @@ const LOW_STACK_TOP: u32 = 0x7ff000;
 /// the fault handler uses it for the SIGSEGV-vs-diagnostic decision.
 #[no_mangle]
 pub extern "C" fn rust_validate_page_fault(
-    fault_addr: u32,
+    fault_addr: u64,
     _error_code: u32,
-    image_base: u32,
-    image_end: u32,
-    heap_start: u32,
-    heap_end: u32,
+    image_base: u64,
+    image_end: u64,
+    heap_start: u64,
+    heap_end: u64,
 ) -> bool {
     // Image: code, rodata, data, bss. image_base is 0 only for an unbuilt task.
     if image_base != 0 && fault_addr >= image_base && fault_addr < image_end {
@@ -100,7 +100,7 @@ pub unsafe extern "C" fn rust_cap_has_rights(cap: *const SafeCapability, require
 }
 
 #[no_mangle]
-pub extern "C" fn rust_get_user_page_protection(_task_id: u32, vaddr: u32) -> u32 {
+pub extern "C" fn rust_get_user_page_protection(_task_id: u32, vaddr: u64) -> u32 {
     // [0xA00000, 0xB00000) is deliberately absent. It used to be excluded because
     // the kernel was linked low and `argon2_scratch` occupied that virtual
     // address; the kernel now runs at KERNEL_VMA, so the reason is different but
@@ -239,7 +239,7 @@ mod tests {
         // heap [0x1000000, 0x1040000). Bounds come from tasks[tid] at the call site.
         let ib = 0x4a0000; let ie = 0x4b6000;
         let hs = 0x1000000; let he = 0x1040000;
-        let v = |a: u32| rust_validate_page_fault(a, 0, ib, ie, hs, he);
+        let v = |a: u64| rust_validate_page_fault(a, 0, ib, ie, hs, he);
 
         // In-region: image, heap, low stack.
         assert!(v(0x4a0000));            // image base
